+++ app.js
@@ -1,6 +1,7 @@
 const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
+require("dotenv").config();
 
 const contactsRouter = require('./routes/api/contacts')
